Move RequeueException to paramore.brighter.commandprocessor
