+++ Brighter/paramore.brighter.commandprocessor.tests/MessageDispatch/TestDoubles/SpyCommandProcessor.cs
@@ -26,6 +26,7 @@ using System;
 using System.Collections.Generic;
 
 using paramore.brighter.commandprocessor;
+using paramore.brighter.commandprocessor.exceptions;
 using paramore.brighter.serviceactivator;
 
 using Raven.Database.Indexing;
