+++ Brighter/paramore.brighter.serviceactivator/paramore.brighter.serviceactivator.csproj
@@ -65,7 +65,6 @@
     <Compile Include="Ports\Handlers\ConfigurationCommandHandler.cs" />
     <Compile Include="Ports\Mappers\ConfigurationCommandMessageMapper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RequeueException.cs" />
     <Compile Include="ServiceActivatorConfiguration\ConnectionElement.cs" />
     <Compile Include="ServiceActivatorConfiguration\ConnectionElements.cs" />
     <Compile Include="ServiceActivatorConfiguration\ConnectionFactory.cs" />
